added rmf_site_mesh + touchups, removed shapes.rs
