+++ Cargo.toml
@@ -28,6 +28,8 @@ rmf_site_editor = { path = "crates/rmf_site_editor" }
 rmf_site_editor_web = { path = "crates/rmf_site_editor_web" }
 rmf_site_animate = {path = "crates/rmf_site_animate"}
 rmf_site_camera = {path = "crates/rmf_site_camera"}
+rmf_site_mesh = {path = "crates/rmf_site_mesh"}
+
 serde_yaml = "0.8.23"
 serde_json = "*"
 ron = "0.10"
