+++ crates/rmf_site_editor/Cargo.toml
@@ -33,6 +33,7 @@ thiserror = {workspace = true}
 rmf_site_format = {workspace = true, features = ["bevy"] }
 rmf_site_animate = {workspace = true}
 rmf_site_camera = {workspace = true}
+rmf_site_mesh = {workspace = true}
 itertools = {workspace = true}
 bitfield = {workspace = true}
 crossbeam-channel = {workspace = true}
