+++ crates/rmf_site_editor/src/site/assets.rs
@@ -15,12 +15,13 @@
  *
 */
 
-use crate::{shapes::*, site::*};
+use crate::{site::*};
 use bevy::{
     asset::embedded_asset,
     math::{primitives, Affine3A},
     prelude::*,
 };
+use rmf_site_mesh::*;
 
 pub(crate) fn add_site_icons(app: &mut App) {
     // Taken from https://github.com/bevyengine/bevy/issues/10377#issuecomment-1858797002
