+++ crates/rmf_site_editor/src/site/drawing.rs
@@ -17,7 +17,6 @@
 
 use crate::{
     interaction::Selectable,
-    shapes::make_flat_rect_mesh,
     site::{
         get_current_workspace_path, Anchor, DefaultFile, FiducialMarker, GlobalDrawingVisibility,
         LayerVisibility, MeasurementMarker, MeasurementSegment, RecencyRank,
@@ -27,6 +26,7 @@ use crate::{
 };
 use bevy::{asset::LoadState, ecs::hierarchy::ChildOf, math::Affine3A, prelude::*};
 use rmf_site_format::{AssetSource, Category, DrawingProperties, PixelsPerMeter, Pose};
+use rmf_site_mesh::*;
 use std::path::PathBuf;
 
 #[derive(Bundle, Debug, Clone)]
