+++ crates/rmf_site_editor/src/site/lift.rs
@@ -16,7 +16,7 @@
 */
 
 use crate::{
-    interaction::Selectable, shapes::*, site::*, CurrentWorkspace, Issue, ValidateWorkspace,
+    interaction::Selectable, site::*, CurrentWorkspace, Issue, ValidateWorkspace,
 };
 use bevy::{
     ecs::{hierarchy::ChildOf, relationship::AncestorIter},
@@ -24,6 +24,7 @@ use bevy::{
     render::primitives::Aabb,
 };
 use rmf_site_format::{Edge, LiftCabin};
+use rmf_site_mesh::*;
 use std::collections::BTreeSet;
 use std::collections::HashMap;
 use uuid::Uuid;
