+++ crates/rmf_site_mesh/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "rmf_site_mesh"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+bevy_ecs = {workspace = true}
+bevy_math = {workspace = true}
+bevy_render = {workspace = true}
+bevy_reflect = {workspace = true}
+bevy_asset = {workspace = true}
+bevy_derive = {workspace = true}
+bevy_color = {workspace = true}
+bevy_mod_outline = {workspace = true}
+
+[dev-dependencies]
+bevy = {workspace = true, features = ["dynamic_linking"]}
+bevy-inspector-egui = {workspace = true}
